ENOENT

+++ so-20-21-ex3_base/so-20-21-ex3_base/fs/main.c
@@ -9,6 +9,7 @@
 #include <sys/types.h>
 #include <sys/socket.h>
 #include <sys/un.h>
+#include <errno.h>
 
 #include "operations.h"
 
@@ -99,8 +100,11 @@ void initServer()
     }
     if (unlink(serverName))
     {
-        perror("initServer: can't unlink given path");
-        exit(EXIT_FAILURE);
+        if(errno != ENOENT){
+            perror("initServer: can't unlink given path");
+            exit(EXIT_FAILURE);
+        }
+        
     }
     addrlen = setSockAddrUn(serverName, &serverAddr);
